#  CCIP-6264 Perf tuning CR configuration for ExecutionStateChange events